Switch to src layout.
The previously used flat layout can lead to some subtle errors with
testing and package distribution. This is due to the fact that the
Python interpreter includes the current working directory as the first
item on the import path, which results in e.g. nox tests being run
against the source tree and not the the installed version of the
package. This is undesirable as tests should be run against the code
that users will install on their machines.