Switch to src layout, and some minor modifications. (#7)
* Add auto-activation of venv via direnv.

* Switch to src layout.

The previously used flat layout can lead to some subtle errors with
testing and package distribution. This is due to the fact that the
Python interpreter includes the current working directory as the first
item on the import path, which results in e.g. nox tests being run
against the source tree and not the the installed version of the
package. This is undesirable as tests should be run against the code
that users will install on their machines.

* Create dev env via nox.

* Switch to from deprecated pkg_resources to importlib.metadata.

---------

Co-authored-by: Roland Kaminski <kaminski@cs.uni-potsdam.de>